+++ src/commandParsing/mathCommandParsing/Product.java
@@ -12,7 +12,7 @@ import drawableobject.DrawableObject;
 public class Product extends TwoInputFloatCommandParser {
 
 	@Override
-	protected float operateOnComponents(List<Float> components,	Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException {
+	protected double operateOnComponents(List<Float> components,	Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException {
 		return components.get(0) * components.get(1);
 	}
