+++ src/commandParsing/mathCommandParsing/Sine.java
@@ -10,8 +10,8 @@ import drawableobject.DrawableObject;
 public class Sine extends OneInputFloatCommandParser {
 
 	@Override
-	protected float operateOnComponents(List<Float> components,	Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException {
-		return (float) Math.sin(components.get(0)/(180/Math.PI));
+	protected double operateOnComponents(List<Float> components,	Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException {
+		return (double) Math.sin(components.get(0)/(180/Math.PI));
 	}
 
 }
